Make the init of ChordPosition public
Because ChordPosition has no init itself; it wil have a member-wise init with internal access level.

That means you can't create a new ChordPosition structs to add or edit chords unless you use a JSONDecoder(). It is not possible to add an 'init' in an extention to ChordPosition.

This commit adds a 'public' init.